requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
google-api-python-client==2.114.0
google-auth==2.27.0
